w3c-reports/12092014/mac-wk2/runned-tests/XMLHttpRequest/send-authentication-basic-setrequestheader-existing-session-expected.txt

Note: this test will only work as expected once per browsing session. Restart browser to re-test.


FAIL XMLHttpRequest: send() - "Basic" authenticated request using setRequestHeader() when there is an existing session assert_equals: expected "90b8b85e-a664-4e09-a683-bc754234fbcb\npass" but got "1cd63cb5-b09d-4b55-bbbf-adf84c30e7c2\nopen-pass"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-authentication-basic-setrequestheader-existing-session.htm:42:26
step@http://localhost:8800/resources/testharness.js:798:30
onreadystatechange@http://localhost:8800/XMLHttpRequest/send-authentication-basic-setrequestheader-existing-session.htm:41:20)
